referral calculation is completed

+++ Pandora.Application/ViewModel/UserPlanReportViewModel.cs
@@ -8,7 +8,7 @@ namespace Pandora.Application.ViewModel
         public string CurrencyName { get; set; }
         public double InvestmentAmount { get; set; }
         public double AccruedProfit { get; set; }
-        public int ProfitPercent { get; set; }
+        public double ProfitPercent { get; set; }
         public bool IsActive { get; set; }
         public string StartDate { get; set; }
         public string EndDate { get; set; }
